Add cpu module and move codes related to dtrace
